Add an install target to the build.
Change the name of the archive library from "lower.a" to "libgoocon.a".
The "lib" prefix enables use with the "-l" linker flag.
I found too much use of "gcl" already in open source,
so I chose the more-but-not-too verbose "goocon" as a base.
Install the library in $(INSTALL_PREFIX)/lib.

Install the headers in $(INSTALL_PREFIX)/include/goocon.
Users of the library will need to add this directory to the include path.

The default $(INSTALL_PREFIX) is /usr/local, but I expect most folks will
simply override it on the make command line.